requirements.txt
google-generativeai>=0.8.0
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0